Refactor: move RequestIDGenerator from core.payments.sci_callback to common.helpers
